[updates-e2e] restructure tests in preparation for new test suite (#18302)
